Move the Iterators into their own modules
